rename files, add jsdoc
